Rename index file
